Move BufferTypeBase and BufferType to separate .h and .cpp files.
