+++ searchlib/src/vespa/searchlib/datastore/array_store.h
@@ -2,6 +2,7 @@
 
 #pragma once
 
+#include "buffer_type.h"
 #include "bufferstate.h"
 #include "datastore.h"
 #include "entryref.h"
